+++ src/filter/mod.rs
@@ -2,10 +2,14 @@ pub use self::size::SizeFilter;
 pub use self::time::TimeFilter;
 
 #[cfg(unix)]
+pub use self::links::LinksFilter;
+#[cfg(unix)]
 pub use self::owner::OwnerFilter;
 
 mod size;
 mod time;
 
+#[cfg(unix)]
+mod links;
 #[cfg(unix)]
 mod owner;